python/mcp_requirements.txt
mcp>=1.0.0
mistralai>=0.0.12
pypdf>=3.17.0
python-pptx>=0.6.21
pillow>=10.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0